+++ src/components/legacy/NavigationBar.stories.js
@@ -2,7 +2,7 @@ import { storiesOf } from "@storybook/vue";
 import { text } from "@storybook/addon-knobs";
 
 import NavigationBar from "./NavigationBar";
-import SchulCloudLogoColor from "@assets/img/logo/logo-image-color.svg";
+import SchulCloudLogoColor from "@assets/img/logo/logo-dBildungscloud.svg";
 
 storiesOf("5 Molecules/NavigationBar", module).add("default", () => ({
 	components: { NavigationBar },