+++ src/components/templates/DefaultWireframe.vue
@@ -4,11 +4,13 @@
 			v-if="breadcrumbs.length"
 			:breadcrumbs="breadcrumbs"
 		></v-custom-breadcrumbs>
+		<div v-else class="breadcrumbs-placeholder"></div>
 		<slot name="header">
 			<h1 class="text-h3">
 				{{ headline }}
 			</h1>
 		</slot>
+		<div class="border"></div>
 		<v-container
 			:class="{
 				'container-max-width': !fullWidth,
@@ -52,6 +54,7 @@ export default {
 };
 </script>
 <style lang="scss" scoped>
+@import "~vuetify/src/styles/styles.sass";
 @import "@styles";
 
 .container.wireframe-container {
@@ -68,5 +71,18 @@ export default {
 .container.container-full-width {
 	max-width: none;
 	margin: 0;
+}
+
+.border {
+	margin-right: calc(-1 * var(--space-lg));
+	margin-bottom: var(--space-xl);
+	margin-left: calc(-1 * var(--space-lg));
+	border-bottom: 2px solid rgba(0, 0, 0, 0.12);
+}
+
+@media #{map-get($display-breakpoints, 'sm-and-up')} {
+	.breadcrumbs-placeholder {
+		height: 24px;
+	}
 }
 </style>